+++ pypeit/core/flux_calib.py
@@ -601,7 +601,8 @@ def find_standard(specobj_list):
 
 
 def sensfunc(wave, counts, counts_ivar, counts_mask, exptime, airmass, std_dict, longitude, latitude, extinctfilepar, ech_orders=None,
-             mask_abs_lines=True, polyorder=4, balm_mask_wid=10.0, nresln=20., resolution=3000.,
+             mask_hydrogen_lines=True, mask_helium_lines=False,
+             polyorder=4, hydrogen_mask_wid=10.0, nresln=20., resolution=3000.,
              trans_thresh=0.9,polycorrect=True, polyfunc=False, debug=False):
     """
     Function to generate the sensitivity function. This function fits
@@ -636,10 +637,12 @@ def sensfunc(wave, counts, counts_ivar, counts_mask, exptime, airmass, std_dict,
         ech_orders (int `numpy.ndarray`_):
             If passed the echelle orders will be added to the meta_table. ech_orders must be a numpy array of integers
             with the shape (norders,) giving the order numbers
-        mask_abs_lines (bool):
-            If True, mask stellar absorption lines before fitting sensitivity function. Default = True
+        mask_hydrogen_lines (bool):
+            If True, mask stellar hydrogen absorption lines before fitting sensitivity function. Default = True
+        mask_helium_lines (bool):
+            If True, mask stellar helium absorption lines before fitting sensitivity function. Default = False
         balm_mask_wid (float):
-            Parameter describing the width of the mask for or stellar absorption lines (i.e. mask_abs_lines=True). A region
+            Parameter describing the width of the mask for or stellar absorption lines (i.e. mask_hydrogen_lines=True). A region
             equal to balm_mask_wid*resln is masked where resln is the estimate for the spectral resolution in pixels
             per resolution element.
         polycorrect (bool):
@@ -682,8 +685,9 @@ def sensfunc(wave, counts, counts_ivar, counts_mask, exptime, airmass, std_dict,
         # Fit the zeropoint
         zeropoint_data[:, iord], zeropoint_data_gpm[:, iord], zeropoint_fit[:, iord], zeropoint_fit_gpm[:, iord], =\
             fit_zeropoint(wave_arr[:,iord], Nlam_star, Nlam_star_ivar, gpm_star, std_dict,
-                          mask_abs_lines=mask_abs_lines, polyorder=polyorder,
-                          balm_mask_wid=balm_mask_wid, nresln=nresln, resolution=resolution, trans_thresh=trans_thresh,
+                          mask_hydrogen_lines=mask_hydrogen_lines, mask_helium_lines=mask_helium_lines,
+                          polyorder=polyorder,
+                          hydrogen_mask_wid=hydrogen_mask_wid, nresln=nresln, resolution=resolution, trans_thresh=trans_thresh,
                           polycorrect=polycorrect, polyfunc=polyfunc, debug=debug)
         # Calculate the minimum and maximum wavelength for this order
         wave_min[iord] = wave_arr[wave_arr[:,iord] > 1.0, iord].min()
@@ -795,7 +799,7 @@ def get_sensfunc_factor(wave, wave_zp, zeropoint, exptime, tellmodel=None, extin
             msgs.error('You must specify longitude and latitude if we are extinction correcting')
         # Apply Extinction if optical bands
         msgs.info("Applying extinction correction")
-        msgs.warn("Extinction correction applyed only if the spectra covers <10000Ang.")
+        msgs.warn("Extinction correction applied only if the spectra covers <10000Ang.")
         extinct = load_extinction_data(longitude, latitude, extinctfilepar)
         ext_corr = extinction_correction(wave * units.AA, airmass, extinct)
         senstot = sensfunc_obs * ext_corr
@@ -859,7 +863,8 @@ def counts2Nlam(wave, counts, counts_ivar, counts_mask, exptime, airmass, longit
 
 
 def fit_zeropoint(wave, Nlam_star, Nlam_ivar_star, gpm_star, std_dict,
-                  mask_abs_lines=True, polyorder=4, balm_mask_wid=10.0, 
+                  mask_hydrogen_lines=True, mask_helium_lines=False,
+                  polyorder=4, hydrogen_mask_wid=10.0,
                   nresln=20., resolution=3000.,
                   trans_thresh=0.9, polycorrect=True, 
                   polyfunc=False, debug=False):
@@ -881,12 +886,14 @@ def fit_zeropoint(wave, Nlam_star, Nlam_ivar_star, gpm_star, std_dict,
             Good pixel mask for Nlam_star
         std_dict (dict):
             Dictionary containing information about the standard star returned by flux_calib.get_standard_spectrum
-        mask_abs_lines (bool, optional):
-            If True, mask stellar absorption lines before fitting sensitivity function. Default = True
-        balm_mask_wid (float, optional):
-            Parameter describing the width of the mask for or stellar absorption lines (i.e. mask_abs_lines=True). A region
-            equal to balm_mask_wid*resln is masked where resln is the estimate for the spectral resolution in pixels
-            per resolution element.
+        mask_hydrogen_lines (bool, optional):
+            If True, mask stellar hydrogen absorption lines before fitting sensitivity function. Default = True
+        mask_helium_lines (bool, optional):
+            If True, mask stellar helium absorption lines before fitting sensitivity function. Default = False
+        hydrogen_mask_wid (float, optional):
+            Parameter describing the width of the mask for or stellar absorption lines (`i.e.`, ``mask_hydrogen_lines=True``)
+            in Angstroms.  A region equal to ``hydrogen_mask_wid`` on either side of the line center is masked.
+            Default = 10A
         polycorrect (bool, optional):
             Whether you want to interpolate the zeropoint with polynomial in the stellar absortion line regions before
             fitting with the bspline
@@ -932,22 +939,25 @@ def fit_zeropoint(wave, Nlam_star, Nlam_ivar_star, gpm_star, std_dict,
             plt.show()
 
     # Get masks from observed star spectrum. True = Good pixels
-    mask_bad, mask_balm, mask_tell = get_mask(wave, Nlam_star, Nlam_ivar_star, gpm_star, mask_abs_lines=mask_abs_lines,
-                                              mask_telluric=True, balm_mask_wid=balm_mask_wid, trans_thresh=trans_thresh)
+    mask_bad, mask_recomb, mask_tell = get_mask(wave, Nlam_star, Nlam_ivar_star, gpm_star,
+                                              mask_hydrogen_lines=mask_hydrogen_lines,
+                                              mask_helium_lines=mask_helium_lines,
+                                              mask_telluric=True, hydrogen_mask_wid=hydrogen_mask_wid,
+                                              trans_thresh=trans_thresh)
 
     # Get zeropoint
     zeropoint_data, zeropoint_data_gpm, zeropoint_fit, zeropoint_fit_gpm = standard_zeropoint(
-        wave, Nlam_star, Nlam_ivar_star, mask_bad, flux_true, mask_balm=mask_balm,
+        wave, Nlam_star, Nlam_ivar_star, mask_bad, flux_true, mask_recomb=mask_recomb,
         mask_tell=mask_tell, maxiter=35, upper=3, lower=3, polyorder=polyorder,
-        balm_mask_wid=balm_mask_wid, nresln=nresln, resolution=resolution,
+        balm_mask_wid=hydrogen_mask_wid, nresln=nresln, resolution=resolution,
         polycorrect=polycorrect, polyfunc=polyfunc, debug=debug)
 
     if debug:
         sensfactor = Nlam_to_Flam(wave, zeropoint_fit)
         plt.plot(wave[zeropoint_fit_gpm], flux_true[zeropoint_fit_gpm], color='k',lw=2, label='Reference Star')
         plt.plot(wave[zeropoint_fit_gpm], Nlam_star[zeropoint_fit_gpm]*sensfactor[zeropoint_fit_gpm], color='r', label='Fluxed Observed Star')
-        plt.xlabel(r'Wavelength [$\AA$]')
-        plt.ylabel('Flux [erg/s/cm2/Ang.]')
+        plt.xlabel('Wavelength (Angstroms)')
+        plt.ylabel(r'Flux (erg/s/cm$^2$/$\AA$)')
         plt.legend(fancybox=True, shadow=True)
         plt.show()
 
@@ -955,9 +965,9 @@ def fit_zeropoint(wave, Nlam_star, Nlam_ivar_star, gpm_star, std_dict,
 
 
 def get_mask(wave_star, flux_star, ivar_star, mask_star, 
-             mask_abs_lines=True, mask_telluric=True, balm_mask_wid=10., 
-             trans_thresh=0.9):
-    """
+             mask_hydrogen_lines=True, mask_helium_lines=False,
+             mask_telluric=True, hydrogen_mask_wid=10., trans_thresh=0.9):
+    r"""
     Generate a set of masks from your observed standard spectrum.
     e.g. Balmer absorption
 
@@ -972,12 +982,20 @@ def get_mask(wave_star, flux_star, ivar_star, mask_star,
     mask_star: bool, optional
         whether you need to mask Hydrogen recombination line region. 
         If False, the returned msk_star are all good.
-    mask_tell: bool, optional
-        whether you need to mask telluric region. If False, the returned msk_tell are all good.
-    balm_mask_wid: float
-      in units of angstrom
-      Mask parameter for Balmer absorption. A region equal to
-      balm_mask_wid is masked.
+    mask_hydrogen_lines: bool, optional
+        whether you need to mask hydrogen absorption lines, mask width set
+        by ``hydrogen_mask_wid``
+    mask_helium_lines: bool, optional
+        whether you need to mask hydrogen absorption lines, mask width set
+        to :math:`0.5 \times` ``hydrogen_mask_wid``
+    mask_telluric: bool, optional
+        whether you need to mask telluric region. If False, the returned
+        msk_tell are all good.
+    hydrogen_mask_wid: float, optional
+        in units of angstrom
+        Mask parameter for hydrogen recombination absorption lines. A region
+        equal to ``hydrogen_mask_wid`` on either side of the line center is
+        masked.
     trans_thresh: float, optional
         parameter for selecting telluric regions.
 
@@ -993,7 +1011,7 @@ def get_mask(wave_star, flux_star, ivar_star, mask_star,
 
     # Mask (True = good pixels)
     # mask for recombination lines
-    mask_balm = np.ones_like(flux_star).astype(bool)
+    mask_recomb = np.ones_like(flux_star).astype(bool)
     # mask for telluric regions
     mask_tell = np.ones_like(flux_star).astype(bool)
 
@@ -1010,43 +1028,15 @@ def get_mask(wave_star, flux_star, ivar_star, mask_star,
     atms_cutoff = wave_star <= 3000.0
     mask_bad[atms_cutoff] = False
 
-    # TODO JFH replace with mask_star_lines from telluric.py
-    if mask_abs_lines:
-        # Mask Balmer, Paschen, Brackett, and Pfund recombination lines
-        msgs.info("Masking recombination lines:")
-        # Mask Balmer
-        msgs.info(" Masking Balmer")
-        lines_balm = np.array([3836.4, 3969.6, 3890.1, 4102.8, 4102.8, 4341.6, 4862.7, 5407.0,
-                               6564.6, 8224.8, 8239.2])
-        for line_balm in lines_balm:
-            ibalm = np.abs(wave_star - line_balm) <= balm_mask_wid
-            mask_balm[ibalm] = False
-        # Mask Paschen
-        msgs.info(" Masking Paschen")
-        # air wavelengths from:
-        # https://www.subarutelescope.org/Science/Resources/lines/hi.html
-        lines_pasc = np.array([8203.6, 8440.3, 8469.6, 8504.8, 8547.7, 8600.8, 8667.4, 8752.9,
-                               8865.2, 9017.4, 9229.0, 9546.0, 10049.4, 10938.1,
-                               12818.1, 18751.0])
-        for line_pasc in lines_pasc:
-            ipasc = np.abs(wave_star - line_pasc) <= balm_mask_wid
-            mask_balm[ipasc] = False
-        # Mask Brackett
-        msgs.info(" Masking Brackett")
-        # air wavelengths from:
-        # https://www.subarutelescope.org/Science/Resources/lines/hi.html
-        lines_brac = np.array([14584.0, 18174.0, 19446.0, 21655.0,26252.0, 40512.0])
-        for line_brac in lines_brac:
-            ibrac = np.abs(wave_star - line_brac) <= balm_mask_wid
-            mask_balm[ibrac] = False
-        # Mask Pfund
-        msgs.info(" Masking Pfund")
-        # air wavelengths from:
-        # https://www.subarutelescope.org/Science/Resources/lines/hi.html
-        lines_pfund = np.array([22788.0, 32961.0, 37395.0, 46525.0,74578.0])
-        for line_pfund in lines_pfund:
-            ipfund = np.abs(wave_star - line_pfund) <= balm_mask_wid
-            mask_balm[ipfund] = False
+    if mask_hydrogen_lines:
+        mask_recomb = mask_stellar_hydrogen(
+            wave_star, mask_width=hydrogen_mask_wid, mask_star=mask_recomb
+        )
+
+    if mask_helium_lines:
+        mask_recomb = mask_stellar_helium(
+            wave_star, mask_width=hydrogen_mask_wid / 2.0, mask_star=mask_recomb
+        )
 
     if mask_telluric:
         ## Mask telluric region in the optical
@@ -1087,7 +1077,123 @@ def get_mask(wave_star, flux_star, ivar_star, mask_star,
         else:
             msgs.info('Your spectrum is bluer than 9100A, only optical telluric regions are masked.')
 
-    return mask_bad, mask_balm, mask_tell
+    return mask_bad, mask_recomb, mask_tell
+
+
+def mask_stellar_hydrogen(wave_star, mask_width=10.0, mask_star=None):
+    """
+    Routine to mask stellar hydrogen recombination lines
+
+    ..note ::
+        This function is pulled out separate from :func:`get_mask` because
+        it is used in the ``telluric`` module, independent of the remainder
+        of the functionality in :func:`get_mask`.
+
+    Args:
+        wave_star (`numpy.ndarray`_):
+            Wavelength of the stellar spectrum
+            shape (nspec,) or (nspec, nimgs)
+        mask_width (float, optional):
+            width to mask on either side of each line center in Angstroms
+        mask_star (`numpy.ndarray`_, optional):
+            Incoming star mask to which to add the hydrogen lines
+            (Default: None)
+
+    Returns:
+        `numpy.ndarray`_:  bool mask.
+           same shape as wave_star, True=Good (i.e. does not hit a stellar absorption line)
+    """
+
+    if mask_star is None:
+        mask_star = np.ones_like(wave_star, dtype=bool)
+    # Mask Balmer, Paschen, Brackett, and Pfund recombination lines
+    msgs.info("Masking hydrogen recombination lines")
+
+    # Mask Balmer
+    msgs.info(" Masking Balmer")
+    # Vacuum Wavelengths from NIST (TEB, 2023-02-10)
+    lines_balm = np.array([6564.6, 4862.7, 4341.7, 4102.9,
+                           3971.2, 3890.2, 3836.4])
+    # Extra lines previously in the list, source unknown:
+    #      [5407.0, 8224.8, 8239.2]
+    for line_balm in lines_balm:
+        ibalm = np.abs(wave_star - line_balm) <= mask_width
+        mask_star[ibalm] = False
+
+    # Mask Paschen
+    msgs.info(" Masking Paschen")
+    # Vacuum Wavelengths from NIST (TEB, 2023-02-10)
+    lines_pasc = np.array([18756.4, 12821.6, 10941.2, 10052.6,
+                           9548.8, 9232.2, 9017.8, 8865.3,
+                           8752.9, 8667.4, 8600.8, 8547.7,
+                           8504.8, 8469.6, 8440.3, 8203.6])
+    for line_pasc in lines_pasc:
+        ipasc = np.abs(wave_star - line_pasc) <= mask_width
+        mask_star[ipasc] = False
+
+    # Mask Brackett
+    msgs.info(" Masking Brackett")
+    # Vacuum Wavelengths from NIST (TEB, 2023-02-10)
+    lines_brac = np.array([40522.8, 26258.7, 21661.2, 19446.0,
+                           18179.2, 17366.9, 14584.0])
+    for line_brac in lines_brac:
+        ibrac = np.abs(wave_star - line_brac) <= mask_width
+        mask_star[ibrac] = False
+
+    # Mask Pfund
+    msgs.info(" Masking Pfund")
+    # Vacuum Wavelengths from NIST (TEB, 2023-02-10)
+    lines_pfund = np.array([74599.0, 46537.8, 37405.8, 32969.8, 22788.0])
+    for line_pfund in lines_pfund:
+        ipfund = np.abs(wave_star - line_pfund) <= mask_width
+        mask_star[ipfund] = False
+
+    return mask_star
+
+
+def mask_stellar_helium(wave_star, mask_width=5.0, mask_star=None):
+    """
+    Routine to mask stellar helium recombination lines
+
+    ..note ::
+        This function is pulled out separate from :func:`get_mask` because
+        it is used in the ``telluric`` module, independent of the remainder
+        of the functionality in :func:`get_mask`.
+
+    Args:
+        wave_star (`numpy.ndarray`_):
+            Wavelength of the stellar spectrum
+            shape (nspec,) or (nspec, nimgs)
+        mask_width (float, optional):
+            width to mask on either side of each line center in Angstroms
+        mask_star (`numpy.ndarray`_, optional):
+            Incoming star mask to which to add the ionized helium lines
+            (Default: None)
+
+    Returns:
+        `numpy.ndarray`_:  bool mask.
+           same shape as wave_star, True=Good (i.e. does not hit a stellar absorption line)
+    """
+
+    if mask_star is None:
+        mask_star = np.ones_like(wave_star, dtype=bool)
+    # Mask Balmer, Paschen, Brackett, and Pfund recombination lines
+    msgs.info("Masking ionized helium recombination lines")
+
+    # Mask HeII
+    msgs.info(" Masking HeII lines")
+    # Prominent HeII lines not overlapped by hydrogen lines:
+    #    Vacuum wavelengths from Hubeney & Milhas (2015)
+    #    "Theory of Stellar Atmospheres", p. 191.
+    lines_heII = np.array([4687.2,   # 3 -> 4
+                            4542.9,   # 4 -> 9
+                            5413.1,   # 4 -> 7
+                            10126.6]) # 4 -> 5
+    for line_heII in lines_heII:
+        iheII = np.abs(wave_star - line_heII) < mask_width
+        mask_star[iheII] = False
+
+    return mask_star
 
 
 # These are physical limits on the allowed values of the zeropoint in magnitudes
@@ -1251,9 +1357,9 @@ def zeropoint_qa_plot(wave, zeropoint_data, zeropoint_data_gpm, zeropoint_fit, z
     rejmask = zeropoint_data_gpm[wv_gpm] & np.logical_not(zeropoint_fit_gpm[wv_gpm])
     axis.plot(wave[wv_gpm], zeropoint_data[wv_gpm], label='Zeropoint estimated', drawstyle='steps-mid', color='k', alpha=0.7, zorder=5, linewidth=1.0)
     axis.plot(wave[wv_gpm], zeropoint_fit[wv_gpm], label='Zeropoint fit', color='red', linewidth=2.0, zorder=7, alpha=0.7)
-    axis.plot(wave[wv_gpm][rejmask], zeropoint_data[wv_gpm][rejmask], 's', zorder=10, mfc='None', mec='blue', mew=0.7, label='rejected pixels')
+    axis.plot(wave[wv_gpm][rejmask], zeropoint_data[wv_gpm][rejmask], 's', zorder=10, mfc='None', mec='blue', mew=0.7, label='rejected pixels from fit')
     axis.plot(wave[wv_gpm][np.logical_not(zeropoint_data_gpm[wv_gpm])], zeropoint_data[wv_gpm][np.logical_not(zeropoint_data_gpm[wv_gpm])], 'v',
-             zorder=9, mfc='None', mec='orange', mew=0.7, label='originally masked')
+             zorder=11, mfc='None', mec='orange', mew=0.7, label='originally masked')
     med_filt_mask = zeropoint_data_gpm[wv_gpm] & np.isfinite(zeropoint_data[wv_gpm])
     zp_med_filter = utils.fast_running_median(zeropoint_data[wv_gpm][med_filt_mask], 11)
     axis.set_ylim(0.95 * zp_med_filter.min(), 1.05 * zp_med_filter.max())
@@ -1266,7 +1372,7 @@ def zeropoint_qa_plot(wave, zeropoint_data, zeropoint_data_gpm, zeropoint_fit, z
 
 
 
-def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=None, mask_tell=None,
+def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_recomb=None, mask_tell=None,
                        maxiter=35, upper=3.0, lower=3.0, func = 'polynomial', polyorder=5, balm_mask_wid=50.,
                        nresln=20., resolution=2700., polycorrect=True, debug=False, polyfunc=False):
     """
@@ -1284,8 +1390,8 @@ def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=Non
       mask for bad pixels. True is good.
     flam_true : Quantity array
       standard star true flux (erg/s/cm^2/A)
-    mask_balm: `numpy.ndarray`_
-      mask for hydrogen recombination lines. True is good.
+    mask_recomb: `numpy.ndarray`_
+      mask for hydrogen (and/or helium II) recombination lines. True is good.
     mask_tell: `numpy.ndarray`_
       mask for telluric regions. True is good.
     maxiter : integer
@@ -1300,6 +1406,8 @@ def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=Non
       in units of angstrom
       Mask parameter for Balmer absorption. A region equal to
       balm_mask_wid is masked.
+    nresln: integer/float
+      number of resolution elements between breakpoints
     resolution: integer/float.
       spectra resolution
       This paramters should be removed in the future. The resolution should be estimated from spectra directly.
@@ -1324,12 +1432,12 @@ def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=Non
     # check masks
     if mask_tell is None:
         mask_tell = np.ones_like(wave,dtype=bool)
-    if mask_balm is None:
-        mask_balm = np.ones_like(wave, dtype=bool)
+    if mask_recomb is None:
+        mask_recomb = np.ones_like(wave, dtype=bool)
 
     zeropoint_data, zeropoint_data_gpm = compute_zeropoint(wave, Nlam, Nlam_gpm, flam_true)
 
-    zeropoint_fitmask = zeropoint_data_gpm & mask_tell & mask_balm
+    zeropoint_fitmask = zeropoint_data_gpm & mask_tell & mask_recomb
     wave_min = wave[wave > 1.0].min()
     wave_max = wave[wave > 1.0].max()
 
@@ -1352,6 +1460,7 @@ def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=Non
     zeropoint_clean_gpm = zeropoint_data_gpm.copy()
     # Polynomial corrections on Hydrogen Recombination lines
     if (np.sum(zeropoint_fitmask) > 0.5 * len(zeropoint_fitmask)) & polycorrect:
+        msgs.info("Replacing bspline fit with polyfit over Hydrogen Recombination line regions")
         ## Only correct Hydrogen Recombination lines with polyfit in the telluric free region
         balmer_clean = np.zeros_like(wave, dtype=bool)
         # Commented out the bluest recombination lines since they are weak for spectroscopic standard stars.
@@ -1385,6 +1494,12 @@ def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=Non
         msgs.warn("Changing input nresln to fix this")
         nresln = std_res / std_pix
 
+    # Output some helpful information for double-checking input params are correct
+    msgs.test(f" This is the passed-in R: {resolution}")
+    msgs.info(f" This is the standard pixel: {std_pix:.2f} Å")
+    msgs.info(f" This is the standard resolution element: {std_res:.2f} Å")
+    msgs.info(f" Breakpoint spacing: {std_res * nresln:.2f} pixels")
+
     # Fit zeropoint with bspline
     kwargs_bspline = {'bkspace': std_res * nresln}
     kwargs_reject = {'maxrej': 5}
@@ -1393,7 +1508,7 @@ def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=Non
     fullbkpt = init_bspline.breakpoints
 
     # remove masked regions from breakpoints
-    msk_bkpt = interpolate.interp1d(wave, zeropoint_clean_gpm, kind='nearest', fill_value='extrapolate')(fullbkpt)
+    msk_bkpt = interpolate.interp1d(wave, zeropoint_fitmask, kind='nearest', fill_value='extrapolate')(fullbkpt)
     init_breakpoints = fullbkpt[msk_bkpt > 0.999]
 
     # init_breakpoints = fullbkpt
@@ -1407,23 +1522,26 @@ def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=Non
     if debug:
         # Check for calibration
         plt.figure(1)
-        plt.plot(wave, zeropoint_data, drawstyle='steps-mid', color='black', label='Zeropoint Data')
-        plt.plot(wave, zeropoint_bspl, color='cornflowerblue', label='Bspline fit')
+        plt.plot(wave, zeropoint_data, drawstyle='steps-mid', color='black', label='Zeropoint Data', zorder=1)
+        plt.plot(wave, zeropoint_bspl, color='cornflowerblue', label='Bspline fit', linewidth=1.0, zorder=2)
+        plt.plot(wave, zeropoint_poly, color='orchid', label='PolyFit for masking', linewidth=2.0, zorder=0)
         plt.plot(wave[np.invert(zeropoint_fitmask)], zeropoint_data[np.invert(zeropoint_fitmask)], '+', color='red', markersize=5.0,
-                 label='masked zeropoint')
-        plt.plot(wave[np.invert(zeropoint_fitmask)], zeropoint_bspl[np.invert(zeropoint_fitmask)], '+', color='red', markersize=5.0,
-                 label='masked zeropoint_bspl_fit')
-        plt.plot(init_breakpoints, zeropoint_bspl_bkpt, '.', color='green', markersize=4.0, label='breakpoints')
-        plt.plot(init_breakpoints, np.interp(init_breakpoints, wave, zeropoint_data), '.', color='green',
+                 label='masked zeropoint_fitmask', zorder=4)
+        plt.plot(wave[np.invert(zeropoint_fit_gpm)], zeropoint_bspl[np.invert(zeropoint_fit_gpm)], 'x', color='pink', markersize=5.0,
+                 label='masked zeropoint_bspl_fit', zorder=3)
+        plt.plot(init_breakpoints, zeropoint_bspl_bkpt, '.', color='cyan', markersize=8.0, label='breakpoints', zorder=10)
+        plt.plot(init_breakpoints, np.interp(init_breakpoints, wave, zeropoint_data), '.', color='green', zorder=15,
                  markersize=4.0,
                  label='data interpolated onto breakpoints')
         plt.plot(wave, 1.0 / np.sqrt(zeropoint_ivar), color='orange', label='sigma used for fits')
         plt.legend()
-        plt.xlabel('Wavelength [ang]')
+        plt.xlabel('Wavelength (Angstroms)')
+        plt.ylabel('Zeropoint (AB mag)')
         med_filt_mask = zeropoint_data_gpm & np.isfinite(zeropoint_data)
         zp_med_filter = utils.fast_running_median(zeropoint_data[med_filt_mask], 11)
         plt.ylim(0.95 * zp_med_filter.min(), 1.05 * zp_med_filter.max())
         plt.title('Bspline fit')
+        plt.legend(fancybox=True, shadow=True)
         plt.show()
 
     if ((np.sum(zeropoint_fitmask) > 0.5 * len(zeropoint_fitmask)) & polycorrect):
@@ -1442,7 +1560,8 @@ def standard_zeropoint(wave, Nlam, Nlam_ivar, Nlam_gpm, flam_true, mask_balm=Non
 
 
     # TODO Should we return the bspline fitmask here?
-    return zeropoint_data, zeropoint_data_gpm, zeropoint_fit, zeropoint_fit_gpm
+    # TODO Shouldn't we return `zeropoint_fitmask` INSTEAD of `zeropoint_data_gpm`? (TPEB, 2/16/23)
+    return zeropoint_data, zeropoint_fitmask, zeropoint_fit, zeropoint_fit_gpm
 
 
 def load_filter_file(filter):